+++ src/places/containers/Places.css
@@ -3,7 +3,7 @@
   margin: 0 auto;
   padding: 1rem;
   width: 90%;
-  max-width: 40rem;
+  max-width: 70vw;
   box-shadow: 0 2px 8px rgba(0, 0, 0, 0.26);
   border-radius: 6px;
   background: white;
